Started Interactive Example for Encyclopedia entry 1.
Entry 1: the <i> element.
Enabled basic events that allow user to select a portion of the text and by pressing 'Ctrl + i' to wrap the
selected text in <i> tags.
This works well while the selected text is before any other text that is displayed in italics. The existence of
<i> tags is messing up the indexes for selection.
This can be solved by using an array or object that will track what parts of the string are modified, and to work
on the innerText instead on innerHTML.